Module_5-Data_Platforms/zoomcamp_ai/pipeline/assets/ingestion/requirements.txt
# TODO: Add Python dependencies required by the python ingestion asset.
# - Keep this intentionally minimal.
# - Pin versions for reproducibility.
# Docs: https://getbruin.com/docs/bruin/assets/python
#
# Example:
# pandas==2.2.0
# requests==2.31.0
pandas==2.2.3
requests==2.32.3
pyarrow==18.1.0
python-dateutil==2.9.0.post0
